Update structure to dao/service/controller layers + config Create Entity Employee(dao+service+controller) Register employee completes just as a standalone part, maybe cooperation with other entities are needed(e.g. one HR can only create employees of his/her own corporation and thus no need to specify the company name of the added employee.
Browse all employees are available.
Plan to finish conditional queries of employees in the weekend, but this may need some interaction with HRs or other modules.

Some abysmal point......
sometimes login will happen to be directed to /error but no error message shown in the log
UNABLE TO FIND THE REASON
OBNOXIOUS